Rework empty string passed to RegCreateKeyEx
This seems to be a practice that some developers followed in the past, or
maybe just that inexperienced developers followed? I'm not entirely sure.
Microsoft employees, such as Raymond Chen, recommend passing nullptr (0)
for this argument instead of an empty string, but from looking at old
forums I can find people saying to pass an empty string instead.

Additionally, I will judge the length of 6 as being a mistake from IDA,
and will simply let the compiler determine the length at compile-time.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>